refactoring
Controllers added
ctrlWrapper added
Middleware added
validate put out
