+++ helpers/index.js
@@ -1,4 +1,5 @@
 const HttpError = require("./HttpError");
 const changeOutputMessage = require("./changeOutputMessage");
+const ctrlWrapper = require("./ctrlWrapper");
 
-module.exports = { HttpError, changeOutputMessage };
+module.exports = { HttpError, changeOutputMessage, ctrlWrapper };
