+++ models/contacts.json
@@ -58,23 +58,5 @@
     "name": "Alec Howard",
     "email": "Donec.elementum@scelerisquescelerisquedui.net",
     "phone": "(748) 206-2688"
-  },
-  {
-    "id": "4vlfs4EC9wthEM7kZK9-R",
-    "name": "John Conor",
-    "email": "john@777.com",
-    "phone": "(777) 777-77777"
-  },
-  {
-    "id": "v04JOQBtLYI6ht-1P7YdW",
-    "name": "John Conor",
-    "email": "john@777.com",
-    "phone": "(777) 777-5555555555"
-  },
-  {
-    "id": "SHQcaCMgruy92DgSrbzmU",
-    "name": "John Conor",
-    "email": "john@777.com",
-    "phone": "(777) 777-888888"
   }
 ]
