+++ schemas/contacts.js
@@ -1,0 +1,15 @@
+const Joi = require("joi");
+
+const addSchema = Joi.object({
+  name: Joi.string().min(3).max(30).required(),
+  email: Joi.string().email().required(),
+  phone: Joi.string().min(3).max(30).required(),
+});
+
+const updateSchema = Joi.object({
+  name: Joi.string().min(3).max(30),
+  email: Joi.string().email(),
+  phone: Joi.string().min(3).max(30),
+}).or("name", "email", "phone");
+
+module.exports = { addSchema, updateSchema };
